refactor 3x3 and added manga query
