+++ src/graphql/searchManga.graphql
@@ -1,0 +1,26 @@
+query searchManga($id: Int, $page: Int, $perPage: Int, $search: String) {
+  Page(page: $page, perPage: $perPage) {
+    pageInfo {
+      total
+      currentPage
+      lastPage
+      hasNextPage
+      perPage
+    }
+    media(id: $id, type: MANGA, sort: POPULARITY_DESC, search: $search) {
+      id
+      coverImage {
+        extraLarge
+        large
+        medium
+        color
+      }
+      format
+      isAdult
+      title {
+        romaji
+        english
+      }
+    }
+  }
+}
